% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cobalt60.R
\docType{data}
\name{Cobalt60}
\alias{Cobalt60}
\title{Cobalt-60}
\format{A data frame with 11 rows and 9 columns}
\description{
A dataset containing the Cytogenetic result obtained from blood 
samples irradiated with gamma-rays (Cobalt-60).
}
\details{
\itemize{
 \item doses - doses;
 \item cell - total number of cells on every dose;
 \item aberr - number of dicentrics detected in a cell;
 \item X0 - number of cells with no dicentrics;
 \item X1 - number of cells with 1 dicentric;
 \item X2 - number of cells with 2 dicentrics;
 \item X3 - number of cells with 3 dicentrics;
 \item X4 - number of cells with 4 dicentrics;
 \item X5 - number of cells with 5 dicentrics;
}
}
\keyword{dataset}

