% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_curve.R
\name{dose_curve}
\alias{dose_curve}
\title{Dose-Effect Curve}
\usage{
dose_curve(x, level = 0.95, xaxis = NULL, newdat = NULL, yci = NULL,
  eta = 1e-04, ...)
}
\arguments{
\item{x}{output from fit.coef function;}

\item{level}{confidence level for fitted values generated;}

\item{xaxis}{x-axis tick labels;}

\item{newdat}{supply newdata}

\item{yci}{default \code{null}; length 2 numeric value of Y_L and Y_U.}

\item{eta}{precision for Y_L and Y_U intersecting C_U and C_L, respectively.}
}
\description{
Generates calibration curve for the chromosomal aberration.
}

