% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mev4He.R
\docType{data}
\name{Mev4He}
\alias{Mev4He}
\title{20 MeV^4He particles}
\format{A data frame with 9 rows and 11 columns}
\description{
A dataset containing the Cytogenetic result obtained from blood 
samples irradiated with Helium-4 particles.
}
\details{
\itemize{
 \item doses - doses;
 \item cell - total number of cells on every dose;
 \item aberr - number of dicentrics detected in a cell;
 \item X0 - number of cells with no dicentrics;
 \item X1 - number of cells with 1 dicentric;
 \item X2 - number of cells with 2 dicentrics;
 \item X3 - number of cells with 3 dicentrics;
 \item X4 - number of cells with 4 dicentrics;
 \item X5 - number of cells with 5 dicentrics;
 \item X6 - number of cells with 6 dicentrics;
 \item X7 - number of cells with 7 dicentrics;
}
}
\keyword{dataset}

