% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Radica.R
\docType{package}
\name{Radica-package}
\alias{Radica-package}
\title{Radiation Induced Chromosome Aberrations}
\description{
radiation induced chromosome aberrations
}
\author{
Al-Ahmadgaid Bahauddin Asaad <alstated@gmail.com>
Herbert Braselmann <braselm@helmholtz-muenchen.de>
}

